migrate casting to seperate component and disable it for now
